chore(arch): reconcile docs with reality + lift shared sync bits into core
Whole-app architecture review verdict: boundaries hold (core purity, iso/server
split, renderer Bridge-only, agent/), sync adapters are genuine adapters, cloud
is clean. The drift was docs lying about the code + three small duplications.

Shared sync bits → @repo/core (kills 3 cross-package duplications):
- fsSafeStamp(date) in sync/reconcile next to its consumer conflictCopyName;
  desktop nodeStamp + mobile createFsStamp are now one-liners over it.
- parseVaultFile/parseVaultManifest boundary parsers in sync/manifest — one
  decoder for the HTTP client, mobile's persisted base cache, and anyone else
  who meets a manifest as `unknown` (mobile's hand-rolled copy deleted).
- ./sync/testing/in-memory-sync-port subpath export — mobile's engine test
  imports it properly instead of a six-deep relative path.

Lint: packages/core/src/** boundary override in .oxlintrc.json (no node/
electron/react/workspace imports) — the repo's most important invariant now
stated in the same system as every other boundary, not just tsconfig.

Docs (the high-leverage fix — CLAUDE.md steers every agent session and was
materially wrong in 4+ places):
- CLAUDE.md: workspace structure gains core/cloud/mobile; "no server-side API"
  qualified (opt-in vault sync; notes never in a server DB); knowledge/markdown
  attributed to @repo/core with features as the node host shell; a Vault sync
  section (engine + adapters + Worker + off-by-default).
- docs/development.md: cloud/mobile dev-loop pointer; vocabulary.ts path fixed
  (now core); Tests section covers all five workspaces.
- Stale README refs to the retired @repo/pi-driver/@repo/agent-runtime aliases;
  features package description (knowledge/markdown moved); knip core comment
  ("consumers that don't exist yet" — they exist); root package name init →
  inteligir (template leftover).

All gates green: typecheck, lint, test (core 141, cloud 15, features 388,
desktop 351, mobile 16), knip, build.
