+++ docs/development.md
@@ -12,7 +12,9 @@ How to run, verify, and change inteligir. Written for humans and agents alike;
 
 ## The two ways to run the app
 
-Same renderer UI both times; they differ in what backs the Bridge.
+Same renderer UI both times; they differ in what backs the Bridge. (The other
+apps have their own loops: `apps/cloud` — `wrangler dev` + the Workers test
+pool, see its README; `apps/mobile` — Expo, needs a device/simulator.)
 
 ### 1. Browser dev harness — fixture Bridge (fastest loop, no backend)
 
@@ -100,17 +102,25 @@ Type-checks passing isn't feature-correct. Drive the running app:
 `apps/desktop/src/renderer/editor/kits/*-kit.tsx`; add the Base half to `base-kit.ts`
 (kit-parity tests fail on drift); a markdown rule in
 `editor/markdown/md-rules.ts` if the node has bytes; vocabulary allowlist in
-`editor/markdown/vocabulary.ts` for MDX nodes; round-trip fixtures proving
-canonical/idempotent behavior.
+`@repo/core/markdown/vocabulary` (`packages/core/src/markdown/vocabulary.ts`)
+for MDX nodes; round-trip fixtures proving canonical/idempotent behavior.
 
 ## Tests
 
+- `pnpm --filter @repo/core test` — the pure domain: vault-sync engine +
+  reconcile + wire contract, knowledge engine (link graph, search, rename),
+  markdown parse/vocabulary.
 - `pnpm --filter @repo/desktop test` — the renderer: editor pipeline (round-trip
   matrix, adversarial harness, kit parity, corpus classification), combobox,
   knowledge fixtures.
-- `pnpm --filter @repo/features test` — the iso contract (knowledge engine,
-  parsers, schemas) **and** the backend under `src/server` (vault, delegation
-  +snapshots, knowledge manager, handlers, secrets).
+- `pnpm --filter @repo/features test` — the iso contract (parsers, schemas)
+  **and** the backend under `src/server` (vault, delegation +snapshots,
+  knowledge manager, sync adapters, handlers, secrets).
+- `pnpm --filter @repo/cloud test` — the sync Worker against real in-process
+  miniflare (DO + R2 + D1 + Better Auth), incl. the end-to-end sync test that
+  drives @repo/core's engine through the real backend.
+- `pnpm --filter @repo/mobile test` — the Expo sync adapters on node (in-memory
+  fakes; no simulator).
 
 ## Releasing the desktop app
