+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "init",
+  "name": "inteligir",
   "private": true,
   "scripts": {
     "build": "turbo run build",
