+++ packages/core/src/sync/manifest.ts
@@ -1,4 +1,5 @@
-import type { Hash, VaultFile, VaultPath } from "./vault-file";
+import { isValidHash, type Hash, type VaultFile, type VaultPath } from "./vault-file";
+import { isRecord } from "./guards";
 
 // ---------------------------------------------------------------------------
 // Manifests — a vault's file set at a point in time, from two vantage points:
@@ -39,3 +40,40 @@ export type LocalManifest = {
   readonly vaultId: string;
   readonly files: readonly LocalFile[];
 };
+
+// ---------------------------------------------------------------------------
+// Boundary parsers — never trust a wire payload or a persisted cache. One
+// implementation for every consumer that decodes a manifest from `unknown`
+// (the HTTP client, a device's persisted base-manifest file, tests).
+// ---------------------------------------------------------------------------
+
+function isNonNegativeInt(value: unknown): value is number {
+  return typeof value === "number" && Number.isInteger(value) && value >= 0;
+}
+
+/** Decode one `VaultFile` from untrusted data, or null if malformed. */
+export function parseVaultFile(raw: unknown): VaultFile | null {
+  if (!isRecord(raw)) return null;
+  const { path, contentHash, version, size } = raw;
+  if (typeof path !== "string" || path === "") return null;
+  if (typeof contentHash !== "string" || !isValidHash(contentHash)) return null;
+  if (!isNonNegativeInt(version) || !isNonNegativeInt(size)) return null;
+  return { path, contentHash, version, size };
+}
+
+/** Decode a `VaultManifest` from untrusted data, or null if any part is
+ * malformed (all-or-nothing — a partial manifest would corrupt a 3-way diff). */
+export function parseVaultManifest(raw: unknown): VaultManifest | null {
+  if (!isRecord(raw)) return null;
+  const { vaultId, generation, files } = raw;
+  if (typeof vaultId !== "string" || !isNonNegativeInt(generation) || !Array.isArray(files)) {
+    return null;
+  }
+  const parsed: VaultFile[] = [];
+  for (const entry of files) {
+    const file = parseVaultFile(entry);
+    if (!file) return null;
+    parsed.push(file);
+  }
+  return { vaultId, generation, files: parsed };
+}
