+++ packages/features/src/server/README.md
@@ -1,6 +1,6 @@
 # `features/src/server` — the node backend
 
-The platform-agnostic backend (the node half of `@repo/features`, imported as `@repo/features/server/*`): app lifecycle, the agent singleton, vault, knowledge indexes, delegation, executor, voice, and the Bridge handler map. `createHost(platform, options)` (`create-host.ts`) composes it all; the Electron desktop shell (`apps/desktop`) injects a `HostPlatform` (`platform.ts`) for everything OS/shell-shaped — native dialogs, secret cipher (keychain or file-key), notifications, packaged-resource paths — folds `host.handlers` over Electron IPC, and forwards `host.events`. Only one real host runs at a time: `lib/host-lock.ts` is a pidfile under `~/.inteligir`. The UI talks to it only via methods declared in the registry (`@repo/features/ipc-registry`). No electron imports anywhere in `server/` — lint-enforced. `pi-driver/` and `agent-runtime/` are folded in as sibling dirs, reached from `agent/` via their `@repo/*` aliases.
+The platform-agnostic backend (the node half of `@repo/features`, imported as `@repo/features/server/*`): app lifecycle, the agent singleton, vault, knowledge indexes, delegation, executor, voice, and the Bridge handler map. `createHost(platform, options)` (`create-host.ts`) composes it all; the Electron desktop shell (`apps/desktop`) injects a `HostPlatform` (`platform.ts`) for everything OS/shell-shaped — native dialogs, secret cipher (keychain or file-key), notifications, packaged-resource paths — folds `host.handlers` over Electron IPC, and forwards `host.events`. Only one real host runs at a time: `lib/host-lock.ts` is a pidfile under `~/.inteligir`. The UI talks to it only via methods declared in the registry (`@repo/features/ipc-registry`). No electron imports anywhere in `server/` — lint-enforced. `pi/` and `agent-runtime/` are folded in as sibling dirs, reached from `agent/` via `@repo/features/server/pi/*` + `@repo/features/server/agent-runtime/*` (the only server paths agent/ may import — lint-enforced).
 
 ## State machine — three-part split
